backend/requirements.txt
Django==6.0.1
channels[daphne]
djangorestframework
markdown
django-filter
djangorestframework-simplejwt
drf-spectacular
django-cors-headers
django-redis
